Q3/Task 01 Cat-vs-Dog-Flask/requirements.txt
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
h5py==2.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.5
tensorflow==2.4.1
tensorflow-estimator==2.4.0
WTForms==2.3.3
pillow~=8.0.1